do away with App.js and put it in index.js
